add auth

+++ app/Http/Controllers/TasksController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Models\Task;
 
 class TasksController extends Controller
@@ -10,7 +11,11 @@ class TasksController extends Controller
     public function index()
     {
         $tasks = Task::all();
-        return view('tasks.index', compact('tasks'));
+        
+        // return view('tasks.index', compact('tasks'));
+        return view('tasks.index', [
+            'tasks' => $tasks,    
+        ]);
     }
 
     public function create()
@@ -23,13 +28,14 @@ class TasksController extends Controller
         // バリデーション
         $request->validate([
             'content' => 'required',         // 入力必須
-            'status'  => 'required||max:10' // 入力必須、10文字以内
+            'status'  => 'required|max:10|in:"未着手","対応中","完了"', // 入力必須、10文字以内
         ]);
         
         // DBに保存
         $task = new Task;
         $task->content = $request->content;
-        $task->stauts  = $request->status;
+        $task->status  = $request->status;
+        $task->user_id = Auth::user()->id;
         $task->save();
         
         return redirect('/');
@@ -37,39 +43,63 @@ class TasksController extends Controller
 
     public function show(string $id)
     {
+        
+        // 「ログインユーザ =タスク所有者」の時のみ処理を許可する
         $task = Task::findOrFail($id);
-        return view('tasks.show', compact('task'));
+        if(Auth::user()->id == $task->user()->first()->id){
+            return view('tasks.show', compact('task'));
+        }
+        
+        return redirect('/');
     }
 
     public function edit(string $id)
     {
+        // 「ログインユーザ =タスク所有者」の時のみ処理を許可する
         $task = Task::findOrFail($id);
-        
-        return view('tasks.edit', compact('task'));
+        if(Auth::user()->id == $task->user()->first()->id){
+            $task = Task::findOrFail($id);
+            return view('tasks.edit', compact('task'));
+        }
+        return redirect('/');
     }
 
     public function update(Request $request, string $id)
     {
-        // バリデーション
-        $request->validate([
-            'content' => 'required',         // 入力必須
-            'status'  => 'required||max:10' // 入力必須、10文字以内
-        ]);
-        
-        // DBに保存
+        // 「ログインユーザ =タスク所有者」の時のみ処理を許可する
         $task = Task::findOrFail($id);
-        $task->content = $request->content;
-        $task->status  = $request->status;
-        $task->save();
+        if(Auth::user()->id == $task->user()->first()->id){
+            
+            // バリデーション
+            $request->validate([
+                'content' => 'required',         // 入力必須
+                'status'  => 'required|max:10' // 入力必須、10文字以内
+            ]);
+            
+            // DBに保存
+            $task = Task::findOrFail($id);
+            $task->content = $request->content;
+            $task->status  = $request->status;
+            $task->save();
         
-        return redirect()->route('tasks.show', [$task->id]);
+            return redirect()->route('tasks.show', [$task->id]);
+        }
+        
+        return redirect('/');
     }
 
     public function destroy(string $id)
     {
+        // 「ログインユーザ =タスク所有者」の時のみ処理を許可する
         $task = Task::findOrFail($id);
-        $task->delete();
-        
-        return redirect('/');
+        if(Auth::user()->id == $task->user()->first()->id){
+            
+            $task = Task::findOrFail($id);
+            $task->delete();
+            
+            return redirect('/')->with('success', 'Delete Successful');
+            
+        }
+        return redirect('/')->with('Delete Failed!');
     }
 }
